refactor(更新模块): 重构更新功能代码，提取通知服务接口
- 将通知功能提取为独立的INotificationService接口
- 实现WindowsToastNotificationService作为具体通知服务
- 重构UpdateService使用依赖注入方式接收通知服务
- 将UpdateWindow逻辑迁移到UpdateViewModel
- 添加对安装器模式资产选择的支持
- 改进错误处理，增加调试日志输出
