+++ Services/Notifications/INotificationService.cs
@@ -1,0 +1,10 @@
+using WindBoard.Models.Update;
+
+namespace WindBoard.Services.Notifications
+{
+    public interface INotificationService
+    {
+        void ShowUpdateAvailable(UpdateInfo updateInfo);
+    }
+}
+
